Module Command.Auto_complete

Specifications for command-line auto-completion.

type t = Univ_map.t -> part:Base.string -> Base.string Base.list

In addition to the argument prefix, an auto-completion spec has access to any previously parsed arguments in the form of a heterogeneous map into which those arguments may register themselves by providing a Univ_map.Multi.Key using the ~key argument to Arg_type.create.

module For_escape : sig ... end