Module Command.For_telemetry

val normalized_path : Base.unit -> Base.string Base.list Base.option

Returns the command and the list of subcommands. Arguments to the Command.t are not included.

If a subcommand was specified as a unique prefix normalized_path will contain the full subcommand name.

If the entry point to the program was not through Command returns None.

val normalized_args : Base.unit -> Base.string Base.list Base.option

Returns the full list of arguments that were parsed by Command.

If a flag name was specified as a unique prefix normalized_args will contain the full flag name.

If the entry point to the program was not through Command returns None.